refactor: restructure sociodemographic architecture
